Revert "Fix parse_string function in debug-bus command (#695)"
This reverts commit 06c8eb99bc9bba042921bac5375fee1ef927180b.
