+++ ttexalens/cli_commands/burst-read-xy.py
@@ -29,7 +29,6 @@ Examples:
 
 command_metadata = {
     "short": "brxy",
-    "long": "burst-read-xy",
     "type": "low-level",
     "description": __doc__,
     "context": ["limited", "metal"],
