+++ ttexalens/cli_commands/debug-bus.py
@@ -52,7 +52,6 @@ Examples:
 
 command_metadata = {
     "short": "dbus",
-    "long": "debug-bus",
     "type": "low-level",
     "description": __doc__,
     "context": ["limited", "metal"],
@@ -110,7 +109,7 @@ def parse_string(input_string: str) -> list[list[int] | str]:
     # group(3): third number
     # group(4): fourth number (optional, may be None)
     # group(5): signal name (when not in {} format)
-    pattern = r"\{([\dA-Fa-fx]+),([\dA-Fa-fx]+),([\dA-Fa-fx]+)(?:,([\dA-Fa-fx]+))?\}|([A-Za-z_][^,\{\}]*)"
+    pattern = r"\{([\dA-Fa-fx]+),([\dA-Fa-fx]+),([\dA-Fa-fx]+)(?:,([\dA-Fa-fx]+))?\}|([A-Za-z_][A-Za-z0-9_/#]*)"
 
     parsed_result: list[list[int] | str] = []
