+++ ttexalens/cli_commands/device-summary.py
