+++ ttexalens/cli_commands/dump-coverage.py
@@ -23,7 +23,6 @@ Examples:
 
 command_metadata = {
     "short": "cov",
-    "long": "dump-coverage",
     "type": "high-level",
     "description": __doc__,
     "context": "limited",
