+++ ttexalens/cli_commands/dump-gpr.py
@@ -19,7 +19,6 @@ Examples:
 """
 command_metadata = {
     "short": "gpr",
-    "long": "dump-gpr",
     "type": "low-level",
     "description": __doc__,
     "context": ["limited", "metal"],
