+++ ttexalens/cli_commands/pci-raw-write.py
@@ -3,25 +3,26 @@
 # SPDX-License-Identifier: Apache-2.0
 """
 Usage:
-  pcir <addr>
+  pciw <addr> <data>
 
 Arguments:
   addr        Address in PCI BAR to read from.
+  data        Data to write to PCI BAR.
 
 Description:
-  Reads data from PCI BAR at address 'addr'. The mapping between the addresses and the on-chip data is stored within the Tensix TLBs.
+  Writes data to PCI BAR at address 'addr'. The mapping between the addresses and the on-chip data is stored within the Tensix TLBs.
 
 Examples:
-  pcir 0x0
+  pciw 0x0 0x0
 """
 
 from docopt import docopt
 
 from ttexalens.uistate import UIState
 
+
 command_metadata = {
-    "short": "pcir",
-    "long": "pci-raw-read",
+    "short": "pciw",
     "type": "dev",
     "description": __doc__,
     "context": ["limited", "metal"],
@@ -31,6 +32,7 @@ command_metadata = {
 def run(cmd_text, context, ui_state: UIState):
     args = docopt(__doc__, argv=cmd_text.split()[1:])
     addr = int(args["<addr>"], 0)
-    pci_read_result = context.server_ifc.pci_read32_raw(ui_state.current_device_id, addr)
-    print(f"PCI RD [0x{addr:x}]: 0x{pci_read_result:x}")
+    data = int(args["<data>"], 0)
+    pci_write_result = context.server_ifc.pci_write32_raw(ui_state.current_device_id, addr, data)
+    print(f"PCI WR [0x{addr:x}] <- 0x{data:x}")
     return None
