+++ ttexalens/cli_commands/run-elf.py
@@ -23,7 +23,6 @@ from ttexalens.uistate import UIState
 
 command_metadata = {
     "short": "re",
-    "long": "run-elf",
     "type": "high-level",
     "description": __doc__,
     "context": ["limited"],
