+++ ttexalens/cli_commands/tensix-reg.py
@@ -41,7 +41,6 @@ Examples:
 
 command_metadata = {
     "short": "reg",
-    "long": "tensix-reg",
     "type": "low-level",
     "description": __doc__,
     "context": ["limited", "metal"],
