+++ ttexalens/cli_commands/write-xy.py
@@ -23,7 +23,6 @@ Examples:
 
 command_metadata = {
     "short": "wxy",
-    "long": "write-xy",
     "type": "low-level",
     "description": __doc__,
     "context": ["limited", "metal"],
